pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbeats-tutorial"
version = "0.1.0"
description = "Agentbeats Tutorial"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "cybergym",
    "httpx>=0.28.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
debate = [
    "google-genai>=1.36.0",
]
tau2-agent = [
    "litellm>=1.0.0",
]
tau2-evaluator = [
    "nest-asyncio>=1.6.0",
    "tau2 @ git+https://github.com/sierra-research/tau2-bench.git",
]

test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
cybergym-green-agent = [
    "cybergym",
    "litellm>=1.80.7",
]
cybergym-purple-agent = [
    "cybergym>=0.1.0",
    "litellm>=1.80.7",
]

[project.scripts]
agentbeats-run = "agentbeats.run_scenario:main"

[tool.uv]
package = true
dev-dependencies = [
    "mypy>=1.18.1",
]

[tool.uv.sources]
cybergym = { git = "https://github.com/sunblaze-ucb/cybergym.git" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agentbeats"]
